modularized preferences components
