+++ frontend/src/components/preferences/types.d.ts
@@ -1,0 +1,9 @@
+export interface PreferenceRange {
+	start_time: string;
+	end_time: string;
+}
+
+export interface DailyPreference {
+	column: Date;
+	ranges: PreferenceRange[];
+}
